requirements.txt
streamlit>=1.28
snowflake-snowpark-python>=0.9.0
snowflake-connector-python>=2.8.0
pyarrow==18.0.0
python-dotenv>=1.0.0